+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# UNTITLED
+
+* Move tests from Pester to mocha
+* Move from TravisCI to Github Actions
+
 # v0.2.0
 
 * Add Powershell 6.2.0-preview.1